Safer use of Eina_Inlist.
Avoid doing casts on it, use EINA_INLIST to define the struct and
EINA_INLIST_GET() to access members, this way if you misused the types
compiler will be able to warn you.

Move most Eina_Inlist loops to EINA_INLIST_FOREACH(), this way it is
more clear and if we want to change something we just change the
macro.

Needs new code in Eina!!! (fix EINA_INLIST_GET() macro)



git-svn-id: http://svn.enlightenment.org/svn/e@39209 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33