Add basic systray support.
You can now dump stalone and others, just note it's still not complete
and not doing XEMBED, so you'll get garbage on icon background. I plan
to fix this in few days.

Note that I'm not getting SelectionClear events, possibly an issue
with ecore_x that I plan to go through later. Without this event I
cannot change the tray to "disabled" and ask it to reacquire the
selection, thus in case you started another tray, you need to remove
and add systray gadget again.



git-svn-id: http://svn.enlightenment.org/svn/e@39631 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33