systray is now usable, but still ugly.
systray now works quite well, at least in my opinion it is better than
stalonetray already (skype icon does not contain garbage on
background), pidgin is at the same size as other icons and so on.

the ugly bit is about base window having a plain solid color but at
least it is theme dependent and with b&w it matches well with white. I
plan to figure out how to make this window a copy of the shelf area
underneath it and draw in the hope icons will do the right thing.



git-svn-id: http://svn.enlightenment.org/svn/e@39640 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33