* e_fm_operation_registry : Add a new api call: EAPI int e_fm2_op_registry_count(void)  * Add the new simple module fileman_opinfo
The module provide a gadget that show the efm operations in progress. It is just a button
with a dynamic label atm, a better interface to come...

To compile the module you need to add it to configure.ac and src/modules/Makefile.am ... I have those
files too much hacked atm to commit :P

k-s: can you commit the 2 files for me please? just add the new module.

Hope you like it
Dave



git-svn-id: http://svn.enlightenment.org/svn/e@39794 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33