Option to enable desklock after screensaver timeout.
This option will make it possible to lock the screen after some
seconds screensaver started.

It is useful in cases you want a short screensaver and uses screenlock
but often need to unlock it quickly (ie: was talking to someone at the
office, screensaver starts and you would like to quickly avoid
desklock, otherwise need to stop talking, unlock and continue).



git-svn-id: http://svn.enlightenment.org/svn/e@40413 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33