unbreak pager and mouse wheel.
pagers should not rely on events on the edje object as it might get
lots of event replicas if there is a chain of swallows/groups/smarts
and they repeat events.

the new theme exports an e.eventarea part that will get the events and
pager will try to use it, if not fallback to the old behavior.

I cannot test this now, please say if it does not work.



git-svn-id: http://svn.enlightenment.org/svn/e@40616 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33