Massive move of old stuff to OLD/
As discussed in the mail list, the following apps were NOT moved in
the hope they will be maintained:

 * edje_viewer (viktor did the port to elementary already)
 * elicit (someone to pick it, possibly rephorm)
 * MISC/gevas* (still maintained, sorry!)

The following were not moved in this batch, but will as soon as we
finish the python bindings and move some code to use ethumb:
 * epsilon
 * BINDINGS/python/python-epsilon

Ephoto will be there until okra adds his new codebase.

All the others were moved to OLD. Note they were not deleted, we're
just assuming they are not maintained and nobody is willing to care
about them. If you wish to step as maintainer, just move stuff back to
trunk or MISC or PROTO and happy hacking.




git-svn-id: http://svn.enlightenment.org/svn/e@41086 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33