Canvas.engine_info
Export and call from test so we end calling engine setup. Engine_Info
is handled as an opaque pointer (integer) and is meant to be forwarded
to extensions in C that would cast that to proper pointer and
configure engines.

This was causing an error with image test, cache was NULL since
evas_common_image_init() was called from eng_setup.



git-svn-id: http://svn.enlightenment.org/svn/e@41340 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33