oops, fix case of timers never being fired while we have idlers.
To reach this case, have a timer that would not be fired on
_ecore_main_loop_iterate_internal(), for example it's not ready yet
(just_added==1), system would get into this inner loop and would never
stop, since there is timer expired now (next_time == 0.0), if we go to
start_loop it would just get into the same loop, not dispatching and
timers.

Python test 04-idler.py triggered that problem.




git-svn-id: http://svn.enlightenment.org/svn/e@41342 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33