fix #397 : the last slide in the slideshow mode was at a wrong position
fix #350 : add a smart object.
Eyelight can be use without the smart object but using it is a good idea.

As Eyelight is a smart object, now we can add a presentation in a presentation.

Remove all warnings

fix some bugs

Remove the pdf support. The pdf was just a list of screenshot of each slide. That's was a bad idea, the pdf could not be resize without losing a lot of quality. We need something better to create a pdf from an evas_object.




git-svn-id: http://svn.enlightenment.org/svn/e@41747 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33