Welcome Editje, a nice tool to graphically edit edje files.
Samsung Electronics just allowed ProFUSION to release Editje as
LGPL-3. They are silently helping improve EFL and Editje is the first
public product of their help.

The software is still in early stages, but already does nice. Have a
try! For that you'll need up to date python bindings
(BINDINGS/python/python-elementary, edje, evas and ecore) and EFL.

Something you'll notice is that Editje does not try to match the
hand-written Edje file techniques, unlike edje_editor and
others. Instead, it was designed together with Rasterman to build on
top of existing infrastructure but in an easy to design way.

More specifically, while doing hand-written edc files we have handful
of states and complex programs that carefully choose states to change
to, Editje creates one state per animation state. This allows easy
timelines to be build, edited and play. So far, all the parts have
states created for all animation steps, but in future we want to allow
users to select the parts that are used in animations, allowing
parallel animations or multiple-timelines.

Again, Editje is not done and you can expect more to come, including
External Widgets support, to match Glade and QtDesigner to some level.



git-svn-id: http://svn.enlightenment.org/svn/e@43724 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33