maximum log level option.
this will have compilers to completely compile out log statements
using levels greater than the given number.

this is defined in config.h, thus C files should include this before
including Eina.h, as should be the case for all files (IOW: if it does
not work for some file, that file already have a bug).




git-svn-id: http://svn.enlightenment.org/svn/e@46482 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33