elicit 2.0 in svn.
Elicit was rewritten by Brian Mattern (rephorm) to avoid the legacy
libraries. The new version is updated to the libs that will be
released soon, so it gets in replacing the old.

His code is also being maintained in GIT:

    http://github.com/rephorm/elicit/

But I hope rephorm is back to commit things to SVN now :-)




git-svn-id: http://svn.enlightenment.org/svn/e@46534 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33