Allow user to redefine the default logging level used by EINA_LOG_ERR(), EINA_LOG_DBG() and other macros.
This will help us to avoid the cumbersome work of having to define the
local macros. I still find the 3-letter macros handy, but we are not
forced to use them now.

I did a small modification to keep the EINA_LOG_DOMAIN_GLOBAL and use
a different symbol for the define, it is EINA_LOG_DOMAIN_DEFAULT. The
documentation was updated to highlight the 2 usage scenario and care
that must be used.


By: Peter Johnson <tam@hiddenrock.com>



git-svn-id: http://svn.enlightenment.org/svn/e@46700 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33