Move esmart to old.
Users are now reduced to software that is already in need for update
after ecore-data, ecore-txt and others are gone.

It is recommended to port your software to use pure Edje and when
higher level widgets are required, go to Elementary. It should be
possible to use Elementary widgets in plain Evas/Edje applications, if
not report the bug with a test case.

NOTE-1: esmart was unmaintained for long time already;

NOTE-2: esmart was the last user of libtool's lt_dl* wrappers, pointless.



git-svn-id: http://svn.enlightenment.org/svn/e@47057 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33