Fix ecore_events for recursive main loops This fixes the following scenario: 1) An event handler starts another main loop 2) The new main loop processes all the remaining event_handlers of this event and the remaining events 3) New events are added to the events list 4) A new iteration occurs
Prior behavior was that on (4) the events already processed were
triggered again. The code added to ecore_suite shows a test case,
similar to the one that is fixed now for modal dialogs on WebKit-EFL. I
preferred to let the INF() messages in order to be easy to copy, paste
and debug outside of the suite if anyone wants to. When the number of
tests grows more, we might want to separate them in different files.

By: Lucas De Marchi <lucas.demarchi@profusion.mobi>



git-svn-id: http://svn.enlightenment.org/svn/e@49390 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33