From: Peter van de Werken <pwerken-e@A-Eskwadraat.nl>
Attached is a patch that makes the config dialog of the conf_desklock
module work better on small screens.  I moved the advanced dialog
options into a toolbook on the basic dialog.  (I used the Screen Saver
dialog as an example.)
In order to make the wallpaper stuff fit I removed the preview widget.



git-svn-id: http://svn.enlightenment.org/svn/e@49472 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33