Select an animation, change the group, go to parts Bug: If you select an animation and go to another group and then change to Parts Mode, an error occurs.
Resolution: cleaning the context and changing to Parts mode before
changing the group.

This also resolves the bug if you is in the animations mode with some
animation selected, change to another group, change back and then go to
Parts Mode, the part state is being the part state of the part in the
Animations Mode.

By: Helen Fornazier <helen.fornazier@profusion.mobi>



git-svn-id: http://svn.enlightenment.org/svn/e@49628 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33