added two new configuration options for the geometry_manipulation conf plugin:
   int         geometry_auto_move; // GUI
   int         geometry_auto_resize_limit; // GUI

This does influence the configure_request algorithm in the way that the user could disable the window 
geometry move and resize feature that is useful for e.g. Gnome fileselector or gnome-terminal tabs.
There was a use case on the list where this was not nice so I created an option to disable it. If one thinks
a special e.src shouldn't activate this feature by default, please change it.

git-svn-id: http://svn.enlightenment.org/svn/e@50083 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33