Efenniht theme for E17
In recent months, k-s, Marina and Tiago were working on a new theme for E17.

Efenniht was devised to be clean and neutral. Its name (which means equinox)
comes from the chromatic duality that was decided at the very beginning, with
a dark theme (black and orange) and a bright one (white and blue) so that more
people feel comfortable using it. Efenniht uses few animations, discrete
contrasts between shades of gray and fine lines (colored) that outline the
selected elements. 

The icons (shown next to the list) were removed due to their small size that
didn’t allow the user to relate them with the words to which they referred.
The margins of the words in respect to the limits of the boxes/windows were
increased, as well as the space between the words of lists. The two horizontal
lines limit the objects and the inner line serves as a guide for a possible
marking occurring during a mouse over event. In the future we plan to
re-introduce some icons, by means of new E17 methods that allow optional
icons (or forced ones), so we can have them for important/remarkable actions
such as “logout”, “run” and “settings”.

This new theme was started from the Black and White because this reflect the
latest changes to E17. The code was broken into modules, in order to
facilitate the development and future maintenance. The current code is 50%
smaller than the Black and White theme, but this will change later on after
implementing this new theme for less used modules to around 35%, which is
still a sizable reduction.

The images used were made mostly by the designer (congrats Marina) and have
good quality (the sources of Photoshop and Illustrator will be made available
soon). We have given attention to the possibility of scalable icons (see the
folder icon in various sizes). Some changes in E code were necessary to enable
the animations that we want, such as signals of selection for E_Icons, and
allowing themes for some modules, such as Illume toggles.

Some features that are still missing are the frame for iBox, which was not
possible now because it is shared with other modules in which the frame was
not desired, and we also skipped the vertical shelf for now, but help with
this will be really appreciated.

The current focus is to use the same concepts for elementary, so
elementary-based applications may look the same with the E17 Efenniht. We’re
starting the elementary theme, but we’ll need some help to finish it in time
to be considered for upcoming release.

Last but not least, most of the module icons were not done. We’re balancing
if we remove or do them, depends on the help we can get there :-)



git-svn-id: http://svn.enlightenment.org/svn/e@50259 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33