Add some extra functionality to the GUI to allow at some rudimentary channel switching/separation.
Still need to use /join #channel to actually join a channel, but now you
can send messages to the channels you are in.  Don't bother using /nick.  It
will change your nick, but you won't be able to send messages as that nick so
it's pointless.  Given this, I'd ask that noone uses it.

The code is quite ugly, but it's temporary.  Mostly needed some GUI constructs
to help test things better so I slapped this together in a hurry.


git-svn-id: http://svn.enlightenment.org/svn/e@51794 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33