[Eve] Improve session saving support.
Use a timer to periodically save the current state: which windows are open,
which tabs are open in which window, which one is focused, etc.  The
relevant entry in the eet file is updated.  This should provide a somewhat
crash-safe approach, which is way better than the previous version.

The timer is currently set to fire every 15s: testing should be performed to
ensure this won't impact performance too much.  Should this be a problem,
this interval might be a user-customized setting.

git-svn-id: http://svn.enlightenment.org/svn/e@52522 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33