test focus brokenness.
Raster, if possible give it a try as it seems that most of the focus
steal and clear is your code and maybe you have a clue of what is
happening. For instance I have no clue why that global focus_order
counter that always grows.

This test is now exposing various problems, some of them is what bug
us in Editje, some is what eve's url bar, etc:

  - press "Give focus to scrolled entry", then press it again. See
    that the scrolled entry looses its focus, although we're giving it
    focus? This is related to the object loosing focus but not
    detecting it, thus the next elm_widget_focus_steal() will just
    ignore it and say "you already have focus, do nothing!"

  - start pressing "TAB", after a full iteration, the next iteration
    just goes up to the last item. If you keep doing that, at some
    point the TAB will just not work anymore.  If you change PARENT
    from "bx" to "win", then it does not happen. But adding to "bx"
    would be the correct thing to do (more logical structure)




git-svn-id: http://svn.enlightenment.org/svn/e@52709 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33