Panel fixed and is now focusable.
The old theme start show animation on obj show, if you want
disabled by default it doesn't work well. I changed the theme
default to be visible, don't need animation on "show".

And hidden_set was with inverted logic.

The focus highlight is in theme. And the widget react to
Return/Space.

git-svn-id: http://svn.enlightenment.org/svn/e@52814 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33