List with els_scroller
Author:    Helen Fornazier <helen.fornazier@profusion.mobi>

The list use elm_scroller instead els_scroller like the others. This is
a problem, because the scroller is son of the list, so when you click on
an item, the event will be done in the item (an edje object) that will
pass to the box it is inside, the box will pass the event to the
scroller and then to the list. So, with the event arriving the scroller
first than the list, if you press up, the navigation between item will
not happen because the scroller is treating the event first. So a good
solution is to pass the list to use smart scroller intead elm_scroller
(like genlist and gengrid).

git-svn-id: http://svn.enlightenment.org/svn/e@53075 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33