New widget: Hoverlist.
This is one more widget whose idea came from Samsung Electronics.  Its
first name was "context popup". There was no common agreement on a
name for it, so... Trator mode, for now (using hoversel as a pattern).
If you have a better name suggestion for it, just change it -- but
soon, due to releasing plans (in a mailbox near in a jiffy).

While for hoversels you populate slots with thingies to later choose
one of them, which will hover up for you, with hoverlists you get a
single hovering list. It will smartly place this -- unique -- list
onto a fitting place inside its parent widget view. Just call
evas_object_show() on it, anytime you want (like for notifies).



git-svn-id: http://svn.enlightenment.org/svn/e@53644 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33