Little titlebar pre-made layout goes in.
This comes from a contributed widget from Samsung that was basically reproducing the code of elm_layout, so instead it was turned into a simple pre-made layout (could get more love, but once again I'm no artist).
In order to make its use simpler, previous commits introduced the signal hooks for elm_layout, text_set and some macros to make it easier to set icon/end and text for layouts like this one.

Also uglified a bit the test for Layout in order to show it working, but I intend to get it fixed again soon. Unless discomfitor wants to practice his Edje skills with it... just saying.


git-svn-id: http://svn.enlightenment.org/svn/e@53971 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33