Edje_Edit has been taught the basics of Embryo scripting.
It is possible now to call a recompile of the script, which if it doesn't
fail, will also update the running Embryo VM. Saving the object, when opened
from a file compiled with a sufficiently new edje_cc (early this week, I think), will generate the source including the scripts in their right place.
It's still missing a proper report of errors during the script build, but that will come later.


git-svn-id: http://svn.enlightenment.org/svn/e@55160 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33