From: Jihoon Kim <jihoon48.kim@samsung.com> Subject: [E-devel] [PATCH] elm_slider patch for disable state
elm_slider does not support disable state, so I make this patch.

Please review it and upload in svn if it is ok.

(nb - also fixed vertical slider sizing in defaul theme and changed
jihoons changes to slider test to make the vertical sliders in a
horizontal box so they dont get silly long windows)



git-svn-id: http://svn.enlightenment.org/svn/e@55340 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33