From: Jae Hwan Kim <jae.hwan.kim@samsung.com> Subject: Re: [E-devel] [PATCH] els_scroller
els_scroller has some problems.

First, when the function "elm_scroller_page_size_set" is used and the
scroller is bounced,
the signal "scroll,anim,stop" is not called.

Second, the signal "scroll,anim,stop" is called at the time
"ELM_THUMBSCROLL_FRICTION",
even if the animation of scroller end before the time.

So I changed the code "els_scroller.c".
Patch code is the following.



git-svn-id: http://svn.enlightenment.org/svn/e@55341 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33