Always check for valid evas object.
If you try to create a new widget, you must be sure that the parent
is really an evas object.

With the previous implementation it was possible to call an _add
function for an elementary widget with any non-null pointer as parent
eventually causing crashes (like with the elm_box).


git-svn-id: http://svn.enlightenment.org/svn/e@55521 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33