Add a simple script to compile EFL.
This script is intended for newcomers to compile all the main libraries and e17
desktop.

To avoid issues mixing different compiled versions, at each execution the
script will make the instalation in a dir named with the revision being
compiled (/opt/e17-@REV@) and all the user has to do is to export the correct
environment variables located at /opt/e17-@REV@/e17-env.sh.

To export the variables of the last successfully compiled version use the
/opt/e17-env.sh link.



git-svn-id: http://svn.enlightenment.org/svn/e@55612 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33