From: Jeonghyun Yun <jh0506.yun@samsung.com> Subject: Re: [E-devel] [Patch] evas_image_load_bmp small patch
When I tested some bmp files on evas, I found one bmp file is not displayed.
But this bmp file is displayed on windows and linux gimp and etc.
So I checked this bmp file, this file have 208 data per line despite 207
pixel per line!!

I fixed code to skip the data when data is more than width instead of
break. I think this is not bug but evas policy problem.



git-svn-id: http://svn.enlightenment.org/svn/e@57076 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33