From: Jae Hwan Kim <jae.hwan.kim@samsung.com>
If the scroller is stopped and set the freeze push while bounce
animation is running, it will stop bouncing in a moment. Even if the scroller
is set the freeze pop again, the scroller don't finish the bouncing. I think
that it should finish the bouncing when it is set the freeze pop.



git-svn-id: http://svn.enlightenment.org/svn/e@58063 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33