From: Seunggyun Kim <sgyun.kim@samsung.com> Subject: [E-devel] [Patch] elm_genlist - added new feature : genlist reorder mode
I introduce a new feature "reorder mode" of genlist.
Using this feature, user can reoder genlist items dynamically.

[API]
==================================================================
- EAPI Eina_Bool   elm_genlist_item_rename_mode_get(Elm_Genlist_Item *it)
Get the rename mode state of an item. This gets the rename mode state
of an item

- EAPI void   elm_genlist_reorder_mode_set(Evas_Object *obj, Eina_Bool reorder_mode)
Set genlist reorder mode. This enables the item is moved to another
item.
==================================================================

After set this reorder api, if user press an item (long press), the item can
be moved another position.
Then if the user releases the item, the item will be reodered in genlist.
When the user releases the holding item, Elm_Genlist_Item_Class.func.moved
callback is also called. 
If needed, the application can add some logic in that callback function.



git-svn-id: http://svn.enlightenment.org/svn/e@59883 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33