emotion: add a gstreamer sink.
This work was first started by Vincent Torri and Nicolas Aguirre.
I took the necessary time to make it work and included inside
emotion with Samsung sponsorship.

This code is not yet completly clean and fast, but it work and
make it possible to add more speed improvement later. This was
a needed infrastructure work.


git-svn-id: http://svn.enlightenment.org/svn/e@60532 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33