Tests: Added elm_tsuite.
Elm_tsuite is a nice script that runs through a list of elementary
applications, takes pictures of all of them and saves them.
After you save the pictures, you can then change (go to a different svn
revision for example) the code and run it again in compare mode to see
if there are any pixel differences.
This suite assumes the applications it tests have only one window which
was created using elementary (elm_window) among other assumptions - all are
documented in the README file.
This depedency does not mean it's only inteneded for testing elementary
widgets, you can test whatever you want, as long as you render it on a window
created with elementary.

git-svn-id: http://svn.enlightenment.org/svn/e@61289 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33