From: Jaehwan Kim <jae.hwan.kim@samsung.com> Subject: Re: [E-devel] [Patch] The bounce problem in the scroller.
The scroller has a problem when it is scrolled consistently in the end of the edge.
It has an acceleration. If the scrolling occur again before the animation end,?
the scroll-animation (momentum animation) be faster. When it is in the end of the edge,
it does, too. So the content of the scroller disappear because of the acceleration.
The scroller don't need to has the acceleration in the end of edge.

The below is the problem.
http://www.youtube.com/watch?v=Qg7BZrm9EnA

And the below is the scroller which be patched.
http://www.youtube.com/watch?v=lHT6G9WoNSs



git-svn-id: http://svn.enlightenment.org/svn/e@61800 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33