elementary/elm_widget - patched by hyoyoung.chang@samsung.com
[E-devel] [patch] elm_widget - elm_widget_type_unregister

Dear developers.

This patch is adding a api to elm_widget.
In elm_widget, there is an api to register widget type to elementary's own
list.
However it doesn't provide unregister api.
So I made this patch.

Most of case, it's not needed. But in some case, it's useful.
(such as elm objects in DSO, if a DSO unloaded, registered widtype is
invalidated)

Thank you



git-svn-id: http://svn.enlightenment.org/svn/e@62099 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33