And Enna come back to Enlightenment SVN .... in a different way, with a different goal. Enna was intented to be a mediacenter for home theater pc, with support of audio, video and music files. For a while it was hosted on Geexbox server, and it will stay there. But this version is completely different. It's named enna-explorer, the source code is based on enna mediacenter, as it took me a long time to have a good implementation of the virtual file mechanism (in my point of view) and I like my theme (almost). So this version is intented to be a file explorer for phones and embedded devices, to be used with a touchscreen and use EFL of course. For now the theme is phones oriented, and doesn't fit well on tablets. But well it's a first shoot,  so please,  be kind :P
Currently, enna-explorer is able to browse your harddrive and it uses Eio for that. You can remove files selected on the user interface. You can't copy/paste yet but i planned to add this features soon. There is still some issues in eio to have a full user experience, like displaying a progress bar during copying/deleting files. But it's planned.
When you select a file, it takes the default program viewer to display it.
Enna-explorer is able to browse upnp devices, and uses gupnp for this purpose. For now you can't display files of upnp shares.
NFS shares are automatically detected if they are already mounted, but you can't yet browse them.
If your samba shares are mounted with fstab it should also detect them but you can't browse them.

I know there is a lot of people that are using EFL tehcnos on embedded devices , i think about SHR guys, Bodhilinux guys ... Please report bugs, and features you would like to see in it. And patches are always accepted



git-svn-id: http://svn.enlightenment.org/svn/e@62534 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33