clock: provide placement orientation and width-for-height and vice-versa.
If no minimum size is provided in the group, it will allocate the
height (for horizontal shelves) or width (for vertical shelves) and
then calculate the required size of the gadget.

It will also send messages to notify the gadget about it's
orientation, if horizontal, vertical or floating. This way it can
adapt its "fit" and relative-to properly.

NOTE: depends on Edje at revision greater than r62640.



git-svn-id: http://svn.enlightenment.org/svn/e@62641 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33