Add generic player backend (with vlc player).
This generic player backend executes a separate player in another
process. It receives the bytes to be drawn on the emotion object through
a shared memory, and communicates with the player through a pipe, using
the player standard input/output.

The player must communicate with emotion using the defined commands
specified in the Emotion_Generic_Plugin.h. It doesn't need to link
against emotion, just include this file for easier implementation.

This implementation was sponsored by Zodiac Aerospace.

git-svn-id: http://svn.enlightenment.org/svn/e@63062 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33