From: Jaehwan Kim <jae.hwan.kim@samsung.com> Subject: [E-devel] [Patch] change rel1.to to default when using group inherit
When using group inherit, if rel1.to is set in parent group, we cannot change the part name 
of rel1.to to default (the default is that dont have the part name) in child group.
So I changed edje_cc_out.c. After patching, we can change rel1.to to
default by using "". (ex > rel1.to: "";)



git-svn-id: http://svn.enlightenment.org/svn/e@64171 7cbeb6ba-43b4-40fd-8cce-4c39aea84d33